// frontend/src/types/bowler.ts
export type bowler ={
    bowlerId: number;
    bowlerFirstName: string;
    bowlerMiddleInit: string;
    bowlerLastName: string;
    bowlerAddress: string;
    bowlerCity: string;
    bowlerState: string;
    bowlerZip: number;
    bowlerPhoneNumber: string;
    teamID: number;
    team: string;
}